requirements.txt
# =========================
# CORE
# =========================
requests==2.32.3
numpy==1.26.4

# =========================
# PLAYWRIGHT (SCRAPER)
# =========================
playwright==1.44.0

# =========================
# DATA PROCESSING
# =========================
pandas==2.2.2

# =========================
# TEXT PROCESSING
# =========================
regex==2024.5.15

# =========================
# PDF / FILE HANDLING
# =========================
pypdf==4.2.0

# =========================
# UI (DASHBOARD)
# =========================
streamlit==1.36.0

# =========================
# OPTIONAL (BETTER EMBEDDING OPS)
# =========================
scikit-learn==1.5.0

# =========================
# ASYNC SUPPORT (SAFE)
# =========================
aiohttp==3.9.5

# =========================
# LOGGING / DEBUG
# =========================
rich==13.7.1

# =========================
# ENTERPRISE & RETRIES
# =========================
tenacity==8.3.0
python-dotenv==1.0.1

playwright
requests